% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{acres_to_kmsq}
\alias{acres_to_kmsq}
\title{Convert acres to square kilometers.}
\usage{
acres_to_kmsq(acres)
}
\arguments{
\item{acres}{numeric acres to convert.}
}
\value{
kilometers squared.
}
\description{
Convert acres to square kilometers.
}
